% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_irt.R
\name{cog_irt}
\alias{cog_irt}
\title{Fit Item Response Theory Models With or Without a Design Matrix}
\usage{
cog_irt(
  data = NULL,
  model = NULL,
  guessing = NULL,
  contrast_codes = NULL,
  num_conditions = NULL,
  num_contrasts = NULL,
  constraints = NULL,
  key = NULL,
  link = "probit",
  ...
)
}
\arguments{
\item{data}{a matrix of item responses (K by IJ). Rows should contain
dichotomous responses (1 or 0) for the items indexed by each column.}

\item{model}{an IRT model name. The options are "1p" for the one-parameter
model, "2p" for the two parameter model, "3p" for the three-parameter, or
"sdt" for a signal detection-weighted model.}

\item{guessing}{either a single numeric guessing value or a matrix of item
guessing parameters (IJ by 1). This argument is only used when model = '3p'.}

\item{contrast_codes}{either a matrix of experimental structure parameters
(JM by MN) or the name in quotes of a R stats contrast function (i.e.,
"contr.helmert", "contr.poly", "contr.sum", "contr.treatment", or
"contr.SAS"). If using the R stats contrast function items in the data matrix
must be arranged by condition.}

\item{num_conditions}{the number of conditions (required if using the R stats
contrast function or when constraints = TRUE).}

\item{num_contrasts}{the number of contrasts including intercept (required if
using the R stats contrast function or when constraints = TRUE).}

\item{constraints}{either a logical (TRUE or FALSE) indicating that item
parameters should be constrained to be equal over the J conditions or a 1 by
I vector of items that should be constrained to be equal across conditions.}

\item{key}{an item key vector where 1 indicates target and 2 indicates
distractor (IJ). Required when model = 'sdt'.}

\item{link}{the name ("logit" or "probit") of the link function to be used in
the model.}

\item{...}{additional arguments.}
}
\value{
List with elements for all parameters estimated, information values
for all parameters estimated, and the model log-likelihood value.
}
\description{
This function estimates item response theory (IRT) model parameters. Users
can optionally choose to request estimates of person parameters representing
effects of the experimental contrast.
}
\section{Dimensions}{

I = Number of items per condition; J = Number of conditions; K = Number of
examinees; M Number of ability (or trait) dimensions; N Number of contrasts
(should include intercept).
}

\examples{

oneparamfit <- cog_irt(data = ex1$y, model = "1p")
summary(oneparamfit)

twoparamfit <- cog_irt(data = ex1$y, model = "2p")
summary(twoparamfit)

threeparamfit <- cog_irt(data = ex1$y, model = "3p", guessing = .25)
summary(threeparamfit)

lrt(oneparamfit, twoparamfit, threeparamfit)

plot(twoparamfit)
plot(ex1$omega[, 1], twoparamfit$omega1[, 1])
abline(a = 0, b = 1)

twoparamfitconstr <- cog_irt(data = ex4$y, model = '2p', contrast_codes =
                            "contr.treatment", num_conditions = 2,
                            num_contrasts = 2, constraints = TRUE)
summary(twoparamfitconstr)
plot(twoparamfitconstr)

# Alternative way to specify constraints
constraints <- matrix(data = c(1,2, 0, 0, 0, 0, 0, 0, 0, 0), nrow = 1,
                 ncol = ncol(x = ex4$y) / 3)
twoparamfitconstr <- cog_irt(data = ex4$y, model = '2p',
                             contrast_codes = "contr.treatment",
                             num_conditions = 2, num_contrasts = 2,
                             constraints = constraints)

# Real data example
sdirt <- cog_irt(data = sternberg$y, model = "sdt", key = sternberg$key)

}
\references{
Embretson S. E., & Reise S. P. (2000). \emph{Item response theory for
psychologists.} Mahwah, N.J.: L. Erlbaum Associates.
Thomas, M. L., Brown, G. G., Patt, V. M., & Duffy, J. R. (2021). Latent
variable modeling and adaptive testing for experimental cognitive
psychopathology research.  \emph{Educational and Psychological Measurement,
81}(1), 155-181.
}
