% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogirt-package.R
\docType{package}
\name{cogirt-package}
\alias{cogirt}
\alias{cogirt-package}
\title{cogirt: Cognitive Testing Using Item Response Theory}
\description{
Psychometrically analyze latent individual differences related to tasks, interventions, or maturational/aging effects in the context of experimental or longitudinal cognitive research using methods first described by Thomas et al. (2020) \doi{10.1177/0013164420919898}.
}
\author{
\strong{Maintainer}: Michael Thomas \email{michael.l.thomas@colostate.edu}

}
\keyword{internal}
