% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cpt}
\alias{cpt}
\title{CPT Data}
\format{
A list with the following elements:
\describe{
  \item{y}{Matrix of dichotomous responses.}
  \item{key}{Item key vector where 1 indicates target and 2 indicates
  distractor (IJ)}
  \item{condition}{Condition vector indiciting distinct conditions or time
  points.}
}
}
\usage{
cpt
}
\description{
CPT task accuracy data collected from an online experiment. The
condition vector indicates backward mask onset (50, 100, 150, or 200 ms).The
key indicates whether items are targets (1) or distractors (2).
}
\keyword{datasets}
