% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dich_response_model.R
\name{dich_response}
\alias{dich_response}
\title{Dichotomous Response Model}
\usage{
dich_response(y, nu, lambda, gamma, omega, zeta, link = "logit")
}
\arguments{
\item{y}{Matrix of item responses (I * J by N).}

\item{nu}{Matrix of item intercept parameters (I * J by N).}

\item{lambda}{Matrix of item structure parameters (I * J by J * M).}

\item{gamma}{Matrix of experimental structure parameters (J * M by K * M).}

\item{omega}{Subject-level effects of the experimental manipulation (K * M by N).}

\item{zeta}{Condition-level prediction errors (N by J * M).}

\item{link}{Choose between logit or probit link functions.}
}
\value{
yhat = response probability; yhatstar = latent response probability;
loglikelihood = model log-likelihood.
}
\description{
This function produces predictions and log-likelihood values for a dichotmous
response model.
}
\examples{
mod <- dich_response(y = sdirt$y, nu = sdirt$nu, lambda = sdirt$lambda,
                     gamma = sdirt$gamma, omega = sdirt$omega, sdirt$zeta)
}
