% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dich_response_deriv.R
\name{dich_response_deriv}
\alias{dich_response_deriv}
\title{Derivatives and Information for the Dichotomous Response Model}
\usage{
dich_response_deriv(y, nu, lambda, gamma, omega, zeta, omega_mu,
  omega_sigma2, zeta_mu, zeta_sigma2, link = "logit")
}
\description{
This function calculates the matrix of first partial dervitives, matrix of
second partial dervitives, and information matrix for the posterior
distribution with respect to omega The formulas are based on Segall (1996;
2009).
}
\section{A Note About Model Notation}{

The function converts GLLVM notation to the more typical IRT notation used by
Segall (1996) for ease of referencing formulas.
}

\examples{
dich_response_deriv(y = sdirt$y, nu = sdirt$nu, lambda = sdirt$lambda,
                gamma = sdirt$gamma, omega = sdirt$omega, zeta = sdirt$zeta,
                omega_mu = sdirt$omega_mu, omega_sigma2 = sdirt$omega_sigma2,
                zeta_mu = sdirt$zeta_mu, zeta_sigma2 = sdirt$zeta_sigma2)

}
\references{
Segall, D. O. (1996). Multidimensional adaptive testing.
\emph{Psychometrika, 61(2)}, 331-354. https://doi.org/10.1007/BF02294343

Segall, D. O. (2009). Principles of Multidimensional Adaptive Testing. In W.
J. van der Linden & C. A. W. Glas (Eds.), \emph{Elements of Adaptive Testing}
 (pp. 57-75). https://doi.org/10.1007/978-0-387-85461-8_3
}
