% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dich_response_model.R
\name{dich_response_model}
\alias{dich_response_model}
\title{Dichotomous Response Model}
\usage{
dich_response_model(y, nu, lambda, gamma, omega, zeta, link = "logit")
}
\arguments{
\item{y}{Matrix of item responses (N by I * J).}

\item{nu}{Matrix of item intercept parameters (N by I * J).}

\item{lambda}{Matrix of item structure parameters (I * J by J * M).}

\item{gamma}{Matrix of experimental structure parameters (J * M by K * M).}

\item{omega}{Examinee-level effects of the experimental manipulation
(N by K * M).}

\item{zeta}{Condition-level effects nested within examineess (N by J * M).}

\item{link}{Choose between logit or probit link functions.}
}
\value{
p = response probability matrix (N by I * J); yhatstar = latent
response variate matrix (N by I * J); loglikelihood = model log-likelihood
(scalar).
}
\description{
This function calculates predictions and log-likelihood values for a
dichotomous response model framed using generalized latent variable modeling
(GLVM; Skrondal & Rabe-Hesketh, 2004).
}
\section{Dimensions}{

N = Number of examinees; I = Number of items per condition; J = Number of
conditions; M = Number of ability (or trait) dimensions; K = Number of
contrasts.
}

\examples{
mod <- dich_response_model(y = sdirt$y, nu = sdirt$nu, lambda = sdirt$lambda,
                     gamma = sdirt$gamma, omega = sdirt$omega, sdirt$zeta)

}
\references{
Skrondal, A., & Rabe-Hesketh, S. (2004). \emph{Generalized latent variable
modeling: Multilevel, longitudinal, and structural equation models}. Boca
Raton: Chapman & Hall/CRC.
}
