% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flanker}
\alias{flanker}
\title{Flanker Data}
\format{
A list with the following elements:
\describe{
  \item{y}{Matrix of dichotomous responses.}
  \item{condition}{Condition vector indiciting distinct conditions or time
  points.}
}
}
\usage{
flanker
}
\description{
Flanker task accuracy data collected from an online experiment. The
condition vector indicates level of congruency ("congruent, incongruent_part,
incongruent_all, neutral).
}
\keyword{datasets}
