% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_fit.R
\name{item_fit}
\alias{item_fit}
\title{Internal: compute a bundle of model-fit diagnostics}
\usage{
item_fit(
  obj,
  n_groups = 5L,
  min_n = 15L,
  B_ppc = 0L,
  ppc_stat = c("tscore_var", "max_q3", "item_means_rmsd")
)
}
\arguments{
\item{n_groups}{Integer, target number of rest-score groups for S-X2-like check.}

\item{min_n}{Integer, minimum group size to include in S-X2-like check.}

\item{B_ppc}{Integer, number of posterior predictive replications. Use 0 to skip.}

\item{ppc_stat}{Character, which PPC stat to use: "tscore_var", "max_q3",
or "item_means_rmsd".}

\item{object}{An object of class 'cog_irt'.}
}
\value{
A list with elements:
  $item_marginals  data.frame
  $item_sx2        data.frame
  $q3              matrix
  $q3_mod          matrix
  $srmsr_pairs     numeric
  $ppc             list or NULL
}
\description{
Given a cogirt-style fitted objectect, compute several diagnostics in one call:
(1) item-level marginals; (2) S-X2-like conditional item fit;
(3) stable Q3 matrix; (4) modified Q3; (5) global SRMSR from pairwise
proportions; and (optionally) (6) a simple posterior predictive check.
}
\details{
Everything is done inline; no separate helper functions are created.
}
