% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrt.R
\name{lrt}
\alias{lrt}
\title{Method of anova for cogirt S3}
\usage{
lrt(object, ...)
}
\arguments{
\item{object}{An object of class 'cogirt'.}

\item{...}{Additional arguments.}
}
\value{
An object of class "anova".
}
\description{
This function compares fit of models produced by cogirt.
}
