% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmhSC.R
\name{mcmhSC}
\alias{mcmhSC}
\title{MCMH Parameter Estimates for Single Chain}
\usage{
mcmhSC(ch = NULL, y = y, objFun = NULL, estOmega = T, estNu = T,
  estZeta = T, lambda = NULL, gamma = NULL, omega0 = NULL,
  nu0 = NULL, zeta0 = NULL, omega_mu = NULL, omega_sigma2 = NULL,
  nu_mu = NULL, nu_sigma2 = NULL, zeta_mu = NULL,
  zeta_sigma2 = NULL, burn = NULL, thin = NULL, min_tune = NULL,
  tune_int = NULL, max_tune = NULL, niter = NULL, weight = NULL,
  verboseMCMH = F)
}
\description{
This function calculates MCMH parameter estimates for a single chain. The
MCMH method implemented follows Patz and Junker (1999). The approach to
tuning the scale and covariance matrix follows BDA and SAS 9.2 User Guide,
2nd Ed. "The MCMC Procedure: Tuning the Proposal Distribution".
}
\examples{
mcmhSC(ch = NULL, y = sdirt$y, objFun = dich_response_model, estOmega = T,
    estNu = T, estZeta = T, lambda = sdirt$lambda, gamma = sdirt$gamma,
    omega0 = array(data = 0, dim = dim(sdirt$omega)),
    nu0 = array(data = 0, dim = c(ncol(sdirt$nu), 1)),
    zeta0 = array(data = 0, dim = dim(sdirt$zeta)),
    omega_mu = sdirt$omega_mu, omega_sigma2 = sdirt$omega_sigma2,
    nu_mu = matrix(sdirt$nu_mu), nu_sigma2 = matrix(sdirt$nu_sigma2),
    zeta_mu = sdirt$zeta_mu, zeta_sigma2 = sdirt$zeta_sigma2,
    burn=0, thin=10, min_tune=50, tune_int=50, max_tune=1000, niter=2000,
    weight=1/1, verboseMCMH=T)

}
\references{
Patz, R. J., & Junker, B. W. (1999). A Straightforward Approach to Markov
Chain Monte Carlo Methods for Item Response Models. Journal of Educational
and Behavioral Statistics, 24(2), 146.
}
