% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nback}
\alias{nback}
\title{N-Back Data}
\format{
A list with the following elements:
\describe{
  \item{y}{Matrix of dichotomous responses.}
  \item{key}{Item key vector where 1 indicates target and 2 indicates
  distractor (IJ)}
  \item{condition}{Condition vector indiciting distinct conditions or time
  points.}
}
}
\usage{
nback
}
\description{
N-Back task accuracy data collected from an online experiment. The condition
vector indicates working memory load level (1-back, 2-back, 3-back, or
4-back). The key indicates whether items are targets (1) or distractors (2).
}
\keyword{datasets}
