% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr.R
\name{nr}
\alias{nr}
\title{Newton Raphson Parameter Estimates}
\usage{
nr(rda = NULL, tol = 1e-09, max_iter = 100, verbose = TRUE)
}
\description{
This function calculates Newton Raphson parameter estimates. It is used
mostly for testing.
}
