% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_nu.R
\name{nr_nu}
\alias{nr_nu}
\title{Newton Raphson Parameter Estimates for Nu}
\usage{
nr_nu(
  y = NULL,
  omega0 = NULL,
  gamma0 = NULL,
  lambda0 = NULL,
  zeta0 = NULL,
  nu0 = NULL,
  kappa0 = NULL,
  nu_mu = NULL,
  nu_sigma2 = NULL,
  link = "probit",
  obj_fun = NULL,
  tol = 1e-09,
  max_iter = 1000,
  verbose = TRUE
)
}
\description{
This function calculates Newton Raphson parameter estimates.
}
\examples{

nr_nu(y = ex1$y, omega0 = ex1$omega, gamma0 = ex1$gamma,
            lambda0 = ex1$lambda, zeta0 = ex1$zeta, nu0 = ex1$nu,
            kappa0 = ex1$kappa, nu_mu = ex1$nu_mu,
            nu_sigma2 = ex1$nu_sigma2, link  = "probit",
            obj_fun = dich_response_model)

}
\keyword{internal}
