% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_omega.R
\name{nr_omega}
\alias{nr_omega}
\title{Newton Raphson Parameter Estimates for Omega Parameters}
\usage{
nr_omega(
  y = NULL,
  omega0 = NULL,
  gamma0 = NULL,
  lambda0 = NULL,
  zeta0 = NULL,
  nu0 = NULL,
  kappa0 = NULL,
  omega_mu = NULL,
  omega_sigma2 = NULL,
  zeta_mu = NULL,
  zeta_sigma2 = NULL,
  link = "probit",
  obj_fun = NULL,
  tol = 1e-09,
  max_iter = 1000,
  verbose = TRUE
)
}
\description{
This function calculates Newton Raphson parameter estimates.
}
\examples{

nr_omega(y = ex1$y, omega0 = ex1$omega, gamma0 = ex1$gamma,
            lambda0 = ex1$lambda, zeta0 = ex1$zeta, nu0 = ex1$nu,
            kappa0 = ex1$kappa, omega_mu = ex1$omega_mu,
            omega_sigma2 = ex1$omega_sigma2, zeta_mu = ex1$zeta_mu,
            zeta_sigma2 = ex1$zeta_sigma2, link = "probit",
            obj_fun = dich_response_model)

}
\keyword{internal}
