% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cog_cat_sim.R
\name{plot.cog_cat_sim}
\alias{plot.cog_cat_sim}
\title{Method of Plot for Simulated Adaptive Testing Using cogirt S3}
\usage{
\method{plot}{cog_cat_sim}(x, ...)
}
\arguments{
\item{x}{An object of class 'cog_cat_sim'.}

\item{...}{Additional arguments.}
}
\value{
This function returns a base R plot displayed in the graphics device.
It does not return any value to the R environment.
}
\description{
This function produces plots for standard errors for
cog_cat_sim results
}
