% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cog_irt.R
\name{plot.cog_irt}
\alias{plot.cog_irt}
\title{Method of Plot for CogIRT S3}
\usage{
\method{plot}{cog_irt}(x, ...)
}
\arguments{
\item{x}{An x of class 'cog_irt'.}

\item{...}{Additional arguments.}
}
\value{
This function returns a base R plot displayed in the graphics device.
It does not return any value to the R environment.
}
\description{
This function produces plots for parameter estimates produced for various
CogIRT models.
}
