% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plt}
\alias{plt}
\title{PLT Data}
\format{
A list with the following elements:
\describe{
  \item{y}{Matrix of dichotomous responses.}
  \item{targ}{Item targ left vs. right vector (IJ)}
  \item{fdbk}{Item fdbk left vs. right vector (IJ)}
  \item{condition}{Condition vector indiciting distinct conditions or time
  points.}
}
}
\usage{
plt
}
\description{
Probabilistic Learning Task (SOPT) accuracy data collected from an online
experiment. The condition vector indicates feedback consistency (90%, 80%,
70%, 60%). The targ vector indicates which side is the target item. The fdbk
vector indicates which side was rewarded.
}
\keyword{datasets}
