% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q3.R
\name{q3}
\alias{q3}
\title{Yen's Q3}
\usage{
q3(object, num_ppc = 200)
}
\arguments{
\item{object}{An object of class 'cog_irt'.}

\item{num_ppc}{Number of posterior predictive replications. Default 200.}
}
\value{
A list with elements for fit information q3, q3_avg, q3_max,
q3_avg_ppp, q3_max_ppp, q3_pair_ppp
}
\description{
Computes Yen's Q3 (residual correlations) and associated posterior predictive
p-values (PPP) computed via parametric replication.
}
\references{
:
Yen, W. M. (1984). Effects of local item dependence on the fit and equating
performance of the three-parameter logistic model. \emph{Applied
Psychological Measurement, 8}, 125-145.
}
