% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmmh.R
\name{rmmh}
\alias{rmmh}
\title{RMMH Parameter Estimates for Multiple Chains}
\usage{
rmmh(chains = 1, y = y, objFun = NULL, estOmega = T, estNu = T,
  estZeta = T, lambda = lambda, gamma = gamma, omega0 = NULL,
  nu0 = NULL, zeta0 = NULL, omega_mu = NULL, omega_sigma2 = NULL,
  nu_mu = NULL, nu_sigma2 = NULL, zeta_mu = NULL,
  zeta_sigma2 = NULL, burn = NULL, thin = NULL, min_tune = NULL,
  tune_int = NULL, max_tune = NULL, niter = NULL, weight = NULL,
  verboseRMMH = T, maxRMMHiter = 200)
}
\description{
This function calculates RMMH parameter estimates for multiple chains.
}
\examples{
rmmh(chains = 3, y = sdirt$y, objFun = dich_response_model, estOmega = T,
    estNu = T, estZeta = T, lambda = sdirt$lambda, gamma = sdirt$gamma,
    omega0 = array(data = 0, dim = dim(sdirt$omega)),
    nu0 = array(data = 0, dim = c(ncol(sdirt$nu), 1)),
    zeta0 = array(data = 0, dim = dim(sdirt$zeta)),
    omega_mu = sdirt$omega_mu, omega_sigma2 = sdirt$omega_sigma2,
    nu_mu = matrix(sdirt$nu_mu), nu_sigma2 = matrix(sdirt$nu_sigma2),
    zeta_mu = sdirt$zeta_mu, zeta_sigma2 = sdirt$zeta_sigma2,
    burn=0, thin=10, min_tune=50, tune_int=50, max_tune = 1000, niter = 2000,
    weight=1/1,verboseRMMH=T, maxRMMHiter=200)

}
\references{

}
