% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sopt}
\alias{sopt}
\title{SOPT Data}
\format{
A list with the following elements:
\describe{
  \item{y}{Matrix of dichotomous responses.}
  \item{condition}{Condition vector indiciting distinct conditions or time
  points.}
}
}
\usage{
sopt
}
\description{
Self-Ordered Pointing Task (SOPT) accuracy data collected from an online
experiment. The condition vector indicates working memory load level (3, 6,
9, or 12 items).
}
\keyword{datasets}
