% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sternberg}
\alias{sternberg}
\title{Sternberg Data}
\format{
A list with the following elements:
\describe{
  \item{y}{Matrix of dichotomous responses.}
  \item{key}{Item key vector where 1 indicates target and 2 indicates
  distractor (IJ)}
  \item{condition}{Condition vector indiciting distinct conditions or time
  points.}
}
}
\usage{
sternberg
}
\description{
Sternberg task accuracy data collected from an online experiment. The
condition vector indicates working memory load level (2, 4, 6, 8, 10, or 12
items).The key indicates whether items are targets (1) or distractors (2).
}
\keyword{datasets}
