% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cog_cat_sim.R
\name{summary.cog_cat_sim}
\alias{summary.cog_cat_sim}
\title{Method of Summary for cog_cat_sim S3}
\usage{
\method{summary}{cog_cat_sim}(object, ...)
}
\arguments{
\item{object}{An object of class 'cog_cat_sim'.}

\item{...}{Additional arguments.}
}
\value{
This function does not return a value to the R environment. Instead,
it prints a comprehensive summary of the simulated computerized adaptive
testing results to the console. The output includes model name and simulation
settings as well as summary statistics for each parameter of interest. The
function is intended for interactive use.
}
\description{
This function provides summary statistics for simulated computerized adaptive
testing.
}
