% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cog_irt.R
\name{summary.cog_irt}
\alias{summary.cog_irt}
\title{Method of Summary for cog_irt S3}
\usage{
\method{summary}{cog_irt}(object, ...)
}
\arguments{
\item{object}{An object of class 'cog_irt'.}

\item{...}{Additional arguments.}
}
\description{
This function provides summary statistics for parameter estimates produced
for various CogIRT models.
}
\keyword{internal}
