% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog_cat.R
\name{cog_cat}
\alias{cog_cat}
\title{Cognitive Testing Using Computerized Adaptive Testing}
\usage{
cog_cat(rda = NULL, obj_fun = NULL, int_par = NULL)
}
\arguments{
\item{obj_fun}{A function that calculates predictions and log-likelihood
values for the selected model (character).}

\item{int_par}{Intentional parameters. That is, the parameters to optimize
precision (scalar).}

\item{x}{.rda file (or list) containing all objects necessary to run rmmh.}
}
\description{
This function takes an rda file or list with select objects and returns omega
estimates, an information matrix, and the next best list to administer for
computerized adaptive testing. Adapting testing using D-optimality (see
Segall 2009).
}
\examples{
rda = sdirtSS
rda$list <- c(sapply(X = 1:(length(rda$y) / 5), FUN = rep, nrep = 5))
rda$y[which(!rda$list \%in\% c(1))] <- NA
cog_cat(rda = rda, obj_fun = dich_response_model, int_par = 1)

}
\references{
Segall, D. O. (2009). Principles of Multidimensional Adaptive Testing. In W.
J. van der Linden & C. A. W. Glas (Eds.), \emph{Elements of Adaptive Testing}
 (pp. 57-75). https://doi.org/10.1007/978-0-387-85461-8_3
}
