% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CogIRT-package.R
\docType{package}
\name{CogIRT-package}
\alias{CogIRT}
\alias{CogIRT-package}
\title{CogIRT: Cognitive Testing Using Item Response Theory}
\description{
Item response theory (IRT) tools for experimental cognitive testing.
}
\author{
\strong{Maintainer}: Michael Thomas \email{michael.l.thomas@colostate.edu}

}
\keyword{internal}
