% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogirt-package.R
\docType{package}
\name{cogirt-package}
\alias{cogirt}
\alias{cogirt-package}
\title{cogirt: Cognitive Testing Using Item Response Theory}
\description{
The cogirt package is designed to help users psychometrically analyze latent individual differences related to tasks, interventions, or maturational/aging effects in the context of experimental or longitudinal cognitive research. The package provides tools for parameter estimation, model comparison, computerized adaptive testing, and data simulation. The methods are based on approaches first described by Thomas et al. (2020) \doi{10.1177/0013164420919898}.
}
\author{
\strong{Maintainer}: Michael Thomas \email{michael.l.thomas@colostate.edu}

}
\keyword{internal}
