% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CogIRT-package.R
\docType{package}
\name{CogIRT-package}
\alias{CogIRT}
\alias{CogIRT-package}
\title{CogIRT: Cognitive Testing Using Item Response Theory}
\description{
CogIRT is an R package designed to support the application of item response theory (IRT) psychometric models in cognitive science research. It enables users to analyze and improve the measurement of individual differences in hypothesized contrast effects, particularly in the context of tasks, interventions, or maturational/aging effects. The package offers tools for parameter estimation, model comparisons, computerized adaptive testing, and data simulation. The methods are based on work of Thomas et al. (2020) \doi{10.1177/0013164420919898}.
}
\author{
\strong{Maintainer}: Michael Thomas \email{michael.l.thomas@colostate.edu}

}
\keyword{internal}
