% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dich_response_model.R
\name{dich_response_model}
\alias{dich_response_model}
\title{Dichotomous Response Model}
\usage{
dich_response_model(
  y = NULL,
  nu = NULL,
  lambda = NULL,
  gamma = NULL,
  omega = NULL,
  zeta = NULL,
  link = "logit"
)
}
\arguments{
\item{y}{Matrix of item responses (K by IJ).}

\item{nu}{Matrix of item intercept parameters (K by IJ).}

\item{lambda}{Matrix of item structure parameters (IJ by JM).}

\item{gamma}{Matrix of experimental structure parameters (JM by MN).}

\item{omega}{Examinee-level effects of the experimental manipulation (K by
MN).}

\item{zeta}{Condition-level effects of the experimental manipulation (K by
JM).}

\item{link}{Choose between logit or probit link functions.}
}
\value{
p = response probability matrix (K by IJ); yhatstar = latent response
variate matrix (K by IJ); loglikelihood = model log-likelihood (scalar).
}
\description{
This function calculates predictions and log-likelihood values for a
dichotomous response model framed using generalized latent variable modeling
(GLVM; Skrondal & Rabe-Hesketh, 2004).
}
\section{Dimensions}{

I = Number of items per condition; J = Number of conditions; K = Number of
examinees; M Number of ability (or trait) dimensions; N Number of contrasts
(should include intercept).
}

\examples{
mod <- dich_response_model(y = sdirt$y, nu = sdirt$nu, lambda = sdirt$lambda,
                     gamma = sdirt$gamma, omega = sdirt$omega, sdirt$zeta)

}
\references{
Skrondal, A., & Rabe-Hesketh, S. (2004). \emph{Generalized latent variable
modeling: Multilevel, longitudinal, and structural equation models}. Boca
Raton: Chapman & Hall/CRC.
}
