% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\title{Item fit}
\usage{
fit(object, ...)
}
\arguments{
\item{object}{An object of class 'cog_irt'.}
}
\value{
A list with elements for fit information q3, ppp_q3
}
\description{
Computes Yen's Q3 (residual correlations) and associated posterior predictive
p-values (PPP).
}
\references{
:
Yen, W. M. (1984). Effects of local item dependence on the fit and equating
performance of the three-parameter logistic model. \emph{Applied
Psychological Measurement, 8}, 125-145.

Li, T., Xie, C., & Jiao, H. (2017). Assessing fit of alternative
unidimensional polytomous IRT models using posterior predictive model
checking. Psychological Methods, 22(2), 397-408.
}
