% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R, R/mr.R
\name{m2}
\alias{m2}
\title{Compute M2 and M2* for IRT model object via Monte Carlo}
\usage{
m2(obj, use_pairs = TRUE)

m2(obj, use_pairs = TRUE)
}
\arguments{
\item{object}{An object of class 'cog_irt'.}

\item{type}{"M2" or "M2*" (small-sample adjusted)}

\item{n_mc}{umber of Monte Carlo samples (draws) used for integration.}
}
\value{
list with statistic, df, p, RMSEA2, and type

list with statistic, df, p, RMSEA2, and type
}
\description{
Uses Monte Carlo draws from the latent trait distribution to estimate
model-implied 1- and 2-way margins, compares them to observed margins, and
returns M2, df, p-value, and an RMSEA-like index (M2*).

Uses Monte Carlo draws from the latent trait distribution to estimate
model-implied 1- and 2-way margins, compares them to observed margins, and
returns M2, df, p-value, and an RMSEA-like index (M2*).
}
\references{
Cai, L. & Hansen, M. (2013). Limited-information goodness-of-object testing of
hierarchical item factor models. \emph{British Journal of Mathematical and
Statistical Psychology, 66}, 245-276.

Cai, L. & Hansen, M. (2013). Limited-information goodness-of-object testing of
hierarchical item factor models. \emph{British Journal of Mathematical and
Statistical Psychology, 66}, 245-276.
}
