% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2.R
\name{m2}
\alias{m2}
\title{Compute M2 and M2* for IRT model object via Monte Carlo}
\usage{
m2(fit, data = NULL)
}
\description{
Compute M2 and M2* for IRT model object via Monte Carlo
}
\references{
Cai, L. & Hansen, M. (2013). Limited-information goodness-of-object testing of
hierarchical item factor models. \emph{British Journal of Mathematical and
Statistical Psychology, 66}, 245-276.
}
