% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cog_cat_sim.R
\name{plot.cog_cat_sim}
\alias{plot.cog_cat_sim}
\title{Method of Plot for Simulated Adaptive Testing Using CogIRT S3}
\usage{
\method{plot}{cog_cat_sim}(object, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{x}{An object of class 'cog_cat_sim'.}
}
\description{
This function produces plots for standard errors for
cog_cat_sim results
}
\keyword{internal}
