% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cog_irt.R
\name{plot.cog_irt}
\alias{plot.cog_irt}
\title{Method of Plot for CogIRT S3}
\usage{
\method{plot}{cog_irt}(object, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{x}{An object of class 'cog_irt'.}
}
\description{
This function produces plots for parameter estimates produced for various
CogIRT models.
}
\keyword{internal}
