% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cog_cat_sim.R
\name{summary.cog_cat_sim}
\alias{summary.cog_cat_sim}
\title{Method of Summary for cog_cat_sim S3}
\usage{
\method{summary}{cog_cat_sim}(object, ...)
}
\arguments{
\item{object}{An object of class 'cog_cat_sim'.}

\item{...}{Additional arguments.}
}
\description{
This function provides summary statistics for simulated computerized adaptive
testing.
}
