% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cog_irt.R
\name{summary.cog_irt}
\alias{summary.cog_irt}
\title{Method of Summary for cog_irt S3}
\usage{
\method{summary}{cog_irt}(object, ...)
}
\arguments{
\item{object}{An object of class 'cog_irt'.}

\item{...}{Additional arguments.}
}
\value{
This function does not return a value to the R environment. Instead,
it prints a detailed summary of the specified IRT model to the console. The
output includes the type of IRT model (e.g., One-Parameter, Two-Parameter,
etc.), the number of subjects and items in the dataset, the log-likelihood of
the model, and summary statistics (mean, standard deviation, median standard
error, and reliability) for estimated parameters. The function is intended
for interactive use to review the results of the fitted model.
}
\description{
This function provides summary statistics for CogIRT models.
}
